+++ src/main/java/com/strangegrotto/wealthdraft/scenarios/Scenario.java
@@ -5,15 +5,22 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.google.common.base.Preconditions;
 import com.strangegrotto.wealthdraft.validator.DeserializationValidator;
 import org.immutables.value.Value;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
+import javax.annotation.Nonnull;
 import java.util.List;
+import java.util.function.Function;
 
 @Value.Immutable
 @JsonDeserialize(as = ImmutableScenario.class)
 public interface Scenario {
+    Logger log = LoggerFactory.getLogger(Scenario.class);
+
     int getYear();
 
-    @JsonProperty("401kContrib") TradOrRothContribs get401kContrib();
+    @JsonProperty("401kContrib")
+    TradOrRothContribs get401kContrib();
 
     TradOrRothContribs getIraContrib();
 
@@ -25,6 +32,13 @@ public interface Scenario {
     List<Long> getLongTermCapitalGains();
 
     List<Long> getShortTermCapitalGains();
+
+    List<Long> getOrdinaryDividends();
+
+    List<Long> getQualifiedDividends();
+
+    // TODO Replace with a set of more descriptive properties
+    List<Long> getOtherUnearnedIncome();
 
     // TODO There are a whoooole bunch of extra AMT adjustments that need to be added back here
     //  Right now this relies on the user to know this, but ideally this should be handled in the YAML
@@ -42,20 +56,31 @@ public interface Scenario {
 
     @Value.Derived
     default IncomeStreams getIncomeStreams() {
-        long earnedIncome = getEarnedIncome().stream()
-                .reduce(0L, (l, r) -> l + r);
-        long ltcg = getLongTermCapitalGains().stream()
-                .reduce(0L, (l, r) -> l + r);
-        long stcg = getShortTermCapitalGains().stream()
-                .reduce(0L, (l, r) -> l + r);
-        // TODO add interest & qualified/regular dividends
-        long otherUnearnedIncome = 0L;
+        Function<List<Long>, Long> aggregateIncome = list -> list.stream().reduce(
+                0L,
+                (l, r) -> l + r
+        );
+
+        log.debug("Ordinary dividends: {}", getOrdinaryDividends());
+
+        long earnedIncome = aggregateIncome.apply(getEarnedIncome());
+        long ltcg = aggregateIncome.apply(getLongTermCapitalGains());
+        long stcg = aggregateIncome.apply(getShortTermCapitalGains());
+        long otherUnearnedIncome = aggregateIncome.apply(getOtherUnearnedIncome());
+        long ordinaryDividends = aggregateIncome.apply(getOrdinaryDividends());
+        long qualifiedDividends = aggregateIncome.apply(getQualifiedDividends());
+
+        long nonPrefUnearnedIncome = stcg +
+                ordinaryDividends +
+                otherUnearnedIncome;
+
+        long prefUnearnedIncome = ltcg +
+                qualifiedDividends;
 
         return ImmutableIncomeStreams.builder()
                 .earnedIncome(earnedIncome)
-                .longTermCapGains(ltcg)
-                .shortTermCapGains(stcg)
-                .otherUnearnedIncome(otherUnearnedIncome)
+                .nonPreferentialUnearnedIncome(nonPrefUnearnedIncome)
+                .preferentialUnearnedIncome(prefUnearnedIncome)
                 .build();
     }
 }