+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+0.1.5
+----
+* Add "Total Taxes" section
+* Add unearned income section to scenario
+* Corrected NIIT calculation
+* Refactored to use distinctions of `earnedIncome`, `nonPreferentialUnearnedIncome`, and `preferentialUnearnedIncome`
+
 0.1.4
 -----
 * Let scenarios specify which gov constant years they're going to use